Move custom color modes to custom hook
