+++ packages/react-app/src/views/HomeView.jsx
@@ -1,12 +1,12 @@
 import React from "react";
-import { useColorModeValue, Container, Box, Heading, Text } from "@chakra-ui/react";
+import { Container, Box, Heading, Text } from "@chakra-ui/react";
 import ChallengeExpandedCard from "../components/ChallengeExpandedCard";
 // ToDo. Ignoring challenge sequence for now.
 import { challengeInfo } from "../data/challenges";
+import useCustomColorModes from "../hooks/useCustomColorModes";
 
 export default function HomeView() {
-  const primaryFontColor = useColorModeValue("gray.700", "gray.200");
-  const secondaryFontColor = useColorModeValue("gray.600", "gray.400");
+  const { primaryFontColor, secondaryFontColor } = useCustomColorModes();
   return (
     <Container maxW="container.md" centerContent>
       <Text color={primaryFontColor} mb="12" fontSize="xl" textAlign="center">
